res fixed some 64bit warnings in sndsys. res added a paragraph about documentation for inherited members to   the coding guidelines. res augmented csInitializer to generate a default app ID from the   executable name (if possible). Also made csApplicationFramework use   this default ID when the user did not set one.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@22391 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csplugincommon/sndsys/convert.h
@@ -131,7 +131,7 @@ public:
    * @param dest_bitspersample
    * @param dest_frequency
    */
-  int ConvertBuffer(const void *source, int source_len, void *dest, 
+  int ConvertBuffer(const void *source, size_t source_len, void *dest, 
     int dest_channels, int dest_bitspersample, int dest_frequency);
 
   /** 
@@ -152,13 +152,18 @@ public:
     int source_frequency);
 
 protected:
-  bool ReadFullSample8(const void **source, int *source_len, int *sample_buffer);
-  bool ReadFullSample16(const void **source, int *source_len, int *sample_buffer);
-  bool ReadFullSample(const void **source, int *source_len, int *sample_buffer);
+  bool ReadFullSample8(const void **source, size_t *source_len, 
+    int *sample_buffer);
+  bool ReadFullSample16(const void **source, size_t *source_len, 
+    int *sample_buffer);
+  bool ReadFullSample(const void **source, size_t  *source_len, 
+    int *sample_buffer);
   int WriteSample8(int *sample_buffer, void **dest, int dest_channels);
   int WriteSample16(int *sample_buffer, void **dest, int dest_channels);
-  int WriteSample(int *sample_buffer, void **dest, int dest_channels, int dest_bitspersample);
-  int AdvanceSourceSamples(const void **source, int *source_len, int samples_to_advance, int *sample_buffer);
+  int WriteSample(int *sample_buffer, void **dest, int dest_channels, 
+    int dest_bitspersample);
+  size_t AdvanceSourceSamples(const void **source, size_t *source_len, 
+    size_t samples_to_advance, int *sample_buffer);
 
 protected:
   int last_sample[CS_SOUND_ELEMENT_MAX_CHANNELS];
